genai_project/requirements.txt
streamlit
langchain
langchain-community
langchain-openai
chromadb
pypdf
python-dotenv
openai
tiktoken
ragas
datasets
pandas
rapidfuzz
